feat(pageable): new utility class to pageable handle.
